+++ releases.md
@@ -1,5 +1,9 @@
 # Releases
 
+## Unreleased
+
+  - Don't make `Kernel#warn` redirection to `Console.warn` the default behavior, you must `require 'console/warn'` to enable it.
+
 ## v1.28.0
 
   - Add support for `Kernel#warn` redirection to `Console.warn`.